+++ Changelog.md
@@ -1,10 +1,14 @@
 # Next
 
 - `cancelled` in the `Cancellable` protocol has been renamed to `isCancelled`.
+- **Breaking Change** Renamed `StructTarget` to `MultiTarget`.
+- Demo project has been updated with new DemoMultiTarget target, new project
+structure and more.
 
 # 8.0.0-beta.5
 
 - Added _validate_ option in `TargetType`, to allow enabling Alamofire automatic validation on requests.
+- Added `mapString(atKeyPath:)` to `Response`, `SignalProducerProtocol`, and `ObservableType`
 
 # 8.0.0-beta.4